requirements.txt
fastapi
uvicorn
python-dotenv
requests
motor
nest_asyncio
regex